Start parsing state at 1, 0
This indicates no characters were parsed yet. If a parser succeeds at
this position, then no characters were consumed. If a parser fails at
this position, then there was no character to consume that could lead
to the failure, otherwise, the character at which the check failed is
returned for failure.